refactor: Implemented COCOKeypointDataset
